#ENG-90393 Split CreatioClientAdapter into its own file
Per review feedback from @d-krestov on PR #652: keep the
IApplicationClient interface declaration and the CreatioClientAdapter
implementation in separate files (clio/Common/IApplicationClient.cs and
clio/Common/CreatioClientAdapter.cs respectively).

No behavior change. Only `using Creatio.Client` and the unused
`Methods: Protected` empty region moved out of the interface file.

Validated: dotnet test --filter "Category=Unit & Module=Common" (439/439 pass).


AI agents: none
